[CORE] Backport https://github.com/bitcoin/bitcoin/commit/9238ecb41752e097443d0bc117df35ebd4ac932e
